refactoring for more logical code
