+++ app/models/YouTubeLiveStream.scala
@@ -1,15 +1,19 @@
-package lib
+package models
 
 import java.net.URI
 
 import com.google.api.services.youtube.model.LiveStream
 
-case class BasicLiveStream (streamName: String, host: String, applicationName: String)
+case class YouTubeLiveStream (
+  streamName: String,
+  host: String,
+  applicationName: String
+)
 
-object BasicLiveStream  {
+object YouTubeLiveStream {
   def build(stream: LiveStream) = {
     val ingestionInfo = stream.getCdn.getIngestionInfo
     val ingestionAddress = new URI(ingestionInfo.getIngestionAddress)
-    BasicLiveStream (ingestionInfo.getStreamName, ingestionAddress.getHost, ingestionAddress.getPath)
+    YouTubeLiveStream(ingestionInfo.getStreamName, ingestionAddress.getHost, ingestionAddress.getPath)
   }
 }
